requirements.txt
unicodecsv
jinja2

google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib